Refactored USB/Driving functions, created OpenCV references
