+++ .idea/gradle.xml
@@ -10,6 +10,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/OpenCV4Android" />
             <option value="$PROJECT_DIR$/app" />
           </set>
         </option>
