+++ Readme.txt
@@ -1,0 +1,7 @@
+This application uses OpenCV to find the best available path for an autonomous car, and outputs the
+appropriate throttle and steering controls over USB Serial to drive along a track.
+
+Compile Instructions:
+1. Download OpenCV4Android and extract to a path that has no spaces (due to Android NDK)
+2. Import the OpenCV4Android SDK folder via Android Studio
+3.
